Unified QuoteDocument data class
