--- src/Data/QuoteDocument/Pisa/pisaQuoteDocumentDataClass.ts
@@ -1,8 +1,0 @@
-import { provideDataClass } from 'scrivito'
-import { pisaConfig } from '../../pisaClient'
-
-export function pisaQuoteDocumentDataClass() {
-  return provideDataClass('QuoteDocument', {
-    restApi: pisaConfig('quote-document'),
-  })
-}
